+++ src/main/java/process/engine/ProducerBulkEngine.java
@@ -56,7 +56,7 @@ public class ProducerBulkEngine {
         JobQueue jobQueue = this.bulkAction.createJobQueueV1(sourceJob.getJobId(), LocalDateTime.now(), JobStatus.Queue, "Job %s now in the queue.", false);
         this.bulkAction.changeJobLastJobRun(sourceJob.getJobId(), jobQueue.getStartTime());
         this.bulkAction.saveJobAuditLogs(jobQueue.getJobQueueId(), String.format("Job %s now in the queue.", sourceJob.getJobId()));
-        this.bulkAction.sendJobStatusNotification(sourceJob.getJobId().intValue());
+        this.bulkAction.sendJobStatusNotification(sourceJob.getJobId());
     }
 
     /**
@@ -68,9 +68,9 @@ public class ProducerBulkEngine {
         JobQueue jobQueue = this.bulkAction.createJobQueueV1(scheduler.getJobId(), scheduler.getRecurrenceTime(), JobStatus.Skip, "Job %s skip, by user action.", true);
         this.bulkAction.saveJobAuditLogs(jobQueue.getJobQueueId(), String.format("Job %s skip, by user action.", scheduler.getJobId()));
         // if the user fail the job manual need to send the mail
-        this.bulkAction.sendJobStatusNotification(jobQueue.getJobId().intValue());
+        this.bulkAction.sendJobStatusNotification(jobQueue.getJobId());
         if (this.transactionService.findByJobId(jobQueue.getJobId()).get().isSkipJob()) {
-            this.emailMessagesFactory.sendSourceJobEmail(EmailMessagesFactory.getSourceJobQueueDto(jobQueue),JobStatus.Skip);
+            this.emailMessagesFactory.sendSourceJobEmail(EmailMessagesFactory.getSourceJobQueueDto(jobQueue), JobStatus.Skip);
         }
     }
 
@@ -101,7 +101,7 @@ public class ProducerBulkEngine {
                             jobQueue = this.bulkAction.createJobQueue(scheduler.getJobId(), LocalDateTime.now(), JobStatus.Skip, "Job %s skip, already in queue.", true);
                             this.bulkAction.saveJobAuditLogs(jobQueue.getJobQueueId(), String.format("Job %s skip, already in queue.", scheduler.getJobId()));
                             if (this.transactionService.findByJobId(scheduler.getJobId()).get().isSkipJob()) {
-                                this.emailMessagesFactory.sendSourceJobEmail(EmailMessagesFactory.getSourceJobQueueDto(jobQueue),JobStatus.Skip);
+                                this.emailMessagesFactory.sendSourceJobEmail(EmailMessagesFactory.getSourceJobQueueDto(jobQueue), JobStatus.Skip);
                             }
                         } else {
                             this.bulkAction.changeJobStatus(scheduler.getJobId(), JobStatus.Queue);
@@ -111,7 +111,7 @@ public class ProducerBulkEngine {
                         }
                         // update the next run in scheduler
                         this.bulkAction.updateNextScheduler(scheduler);
-                        this.bulkAction.sendJobStatusNotification(scheduler.getJobId().intValue());
+                        this.bulkAction.sendJobStatusNotification(scheduler.getJobId());
                     }
                 });
                 return;
@@ -244,9 +244,9 @@ public class ProducerBulkEngine {
         this.bulkAction.changeJobQueueStatus(jobQueue.getJobQueueId(), JobStatus.Failed);
         this.bulkAction.saveJobAuditLogs(jobQueue.getJobQueueId(), String.format(message, jobQueue.getJobId()));
         this.bulkAction.changeJobQueueEndDate(jobQueue.getJobQueueId(), LocalDateTime.now());
-        this.bulkAction.sendJobStatusNotification(jobQueue.getJobId().intValue());
+        this.bulkAction.sendJobStatusNotification(jobQueue.getJobId());
         if (this.transactionService.findByJobId(jobQueue.getJobId()).get().isFailJob()) {
-            this.emailMessagesFactory.sendSourceJobEmail(EmailMessagesFactory.getSourceJobQueueDto(jobQueue),JobStatus.Failed);
+            this.emailMessagesFactory.sendSourceJobEmail(EmailMessagesFactory.getSourceJobQueueDto(jobQueue), JobStatus.Failed);
         }
     }
